+++ src/main/data/duke.txt
@@ -1,0 +1,4 @@
+T | 1 | homework
+T | 0 | homework
+
+D | 0 | submit code | 12th of September 2019, 12.00pm
