+++ src/main/java/Parser.java
@@ -1,3 +1,9 @@
+import command.*;
+import exception.DukeException;
+import task.Deadline;
+import task.Todo;
+import task.Event;
+
 import java.util.Arrays;
 
 /**
