+++ src/main/java/command/BadCommand.java
@@ -1,3 +1,9 @@
+package command;
+
+import storage.Storage;
+import task.TaskList;
+import ui.Ui;
+
 public class BadCommand extends Command {
 
     @Override
