+++ src/main/java/command/ExitCommand.java
@@ -1,12 +1,19 @@
+package command;
+
+import storage.Storage;
+import task.TaskList;
+import ui.Ui;
+
 /**
- * Represents a command from user to lists current tasks.
+ * Represents the command from user to exit application.
  * Inherits from Command class.
  * @author Zhang Yue Han
  */
-public class ListCommand extends Command {
+
+public class ExitCommand extends Command {
 
     /**
-     * Prints the current task list.
+     * Prints exit message to user and exits the application.
      * @param ui Ui object used to display information to the user
      * @param tasks TaskList object which contains the task array list holding the task info
      * @param storage Storage object which is used to write new task to file
@@ -16,11 +23,16 @@ public class ListCommand extends Command {
         //ask ui to print something
         //ask tasks to store the thing in arraylist
         //ask storage to write to file
-        try {
-            if (tasks.getList().isEmpty()) { throw new DukeException(DukeException.ErrorType.LIST_EMPTY); }
-            ui.showList(tasks);
-        } catch (DukeException e) {
-            e.showError();
-        }
+        ui.goodBye();
+        isExit();
+    }
+
+    /**
+     * Changes the exit boolean to be true.
+     * @return the value of exit as true
+     */
+    @Override
+    public boolean isExit() {
+        return true;
     }
 }
