+++ src/main/java/command/ListCommand.java
@@ -1,12 +1,19 @@
+package command;
+
+import exception.DukeException;
+import storage.Storage;
+import task.TaskList;
+import ui.Ui;
+
 /**
- * Represents the command from user to exit application.
+ * Represents a command from user to lists current tasks.
  * Inherits from Command class.
  * @author Zhang Yue Han
  */
-public class ExitCommand extends Command {
+public class ListCommand extends Command {
 
     /**
-     * Prints exit message to user and exits the application.
+     * Prints the current task list.
      * @param ui Ui object used to display information to the user
      * @param tasks TaskList object which contains the task array list holding the task info
      * @param storage Storage object which is used to write new task to file
@@ -16,16 +23,11 @@ public class ExitCommand extends Command {
         //ask ui to print something
         //ask tasks to store the thing in arraylist
         //ask storage to write to file
-        ui.goodBye();
-        isExit();
-    }
-
-    /**
-     * Changes the exit boolean to be true.
-     * @return the value of exit as true
-     */
-    @Override
-    public boolean isExit() {
-        return true;
+        try {
+            if (tasks.getList().isEmpty()) { throw new DukeException(DukeException.ErrorType.LIST_EMPTY); }
+            ui.showList(tasks);
+        } catch (DukeException e) {
+            e.showError();
+        }
     }
 }
