+++ src/main/java/task/Task.java
@@ -1,3 +1,5 @@
+package task;
+
 /**
  * Represents a task object containing task details.
  * @author Zhang Yue Han
